Feat/activity types

See merge request ip34-24bb/ip34-24bb_shakethelake/shakethelakeadminfrontend!30

+++ src/services/event-service.ts
@@ -27,7 +27,8 @@ export const createEvent = async (event: EventDto): Promise<EventDto> => {
 	return response.data;
 };
 
-export const deleteEvent = async (id: number): Promise<void> => {
+export const deleteEvent = async (id?: number): Promise<void> => {
+	if (!id) return;
 	await axios.delete(`${baseUrl}/event/${id}`);
 };
